requirements.txt
asgiref==3.2.7
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.8.2
certifi==2019.3.9
chardet==3.0.4
codecov==2.0.15
colorama==0.4.3
coverage==4.5.3
Django==3.0.6
django-modelcluster==5.0.2
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
html5lib==1.1
idna==2.8
l18n==2018.5
more-itertools==6.0.0
mypy-extensions==0.4.1
path==13.2.0
Pillow==7.2.0
pluggy==0.9.0
py==1.8.0
pytest==4.3.1
pytz==2020.1
requests==2.21.0
six==1.12.0
soupsieve==2.0.1
sqlparse==0.3.1
survey==1.0.0
terminaltables==3.1.0
Unidecode==1.1.1
urllib3==1.24.1
wagtail==2.9.2
webencodings==0.5.1
Willow==1.3
XlsxWriter==1.2.9
